Development changes for release 0.2.5

+++ src/mem.h
@@ -17,14 +17,14 @@
  */
 
 #ifdef __cplusplus
-extern "C" {
+  extern "C" {
 #endif
 
 void MemInit();
 void *xMemAlloc(size_t);
 void xMemFree(void *);
-int xMemGetUsed();
-int xMemGetSize(void *);
+size_t xMemGetUsed();
+size_t xMemGetSize(void *);
 void MemClear();
 
 #ifdef __cplusplus
